pyproject.toml
[project]
name = "scanner-app"
version = "1.0.0"
description = "Scanner Application for HP Deskjet 2130"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
    "reportlab>=4.0.0",
    "pywin32>=305",
    "wmi>=1.5.1",
    "pysimplegui>=5.0.8.3",
    "comtypes>=1.4.12",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
